feat: breaking step components and adding custom hook for password validation
