+++ .chores.md
@@ -1,8 +1,14 @@
 - [x] Change schema to fix validations
 - [x] Refactor to apply Single Responsibility Principle to validation (validation is done on joi only)
+- [ ] Validation on change
+- [ ] Validation on submit
 - [ ] Take validation out from the component.
   - [x] SetNewPassword
   - [ ] Register
+    - [ ] Refactor on change to deal with step three specificity
+    - [ ] Fix Input enter not working
+  - [ ] PasswordPanel
+  - [ ] Login
 - [x] Update check component to accept not highlighted text
 - [x] Split some of these user feedbacks? Like:
   - [x] Must contain at least one special character,
@@ -11,5 +17,13 @@
   - [x] Must contain at least one lower character
 - [x] Pass mismatching password user feedback as the other messages, for consistency
 - [ ] Get rid of toast message
-- [x] Check for confirm password if password changes and confirm is filled, and they do not match anymore
-- [ ] Check for confirm password if password changes, and they now match
+- [x] Tests
+  - [x] Valid Password and Matching Confirmation
+  - [x] Valid Password, Incorrect Confirmation, then Corrects confirmation
+  - [x] Password Update to Match Confirm Password
+  - [x] Short Password with Matching Confirmation
+  - [x] Invalid Password Format, Correct Confirmation
+  - [x] Empty Password Field with Confirmation Filled
+  - [x] Confirm Password Left Blank
+  - [x] Password Change After Successful Match
+- [x] Avoid submission in new password if password or confirm empty, and on error
